+++ src/services/BlockTypes.php
@@ -282,11 +282,19 @@ class BlockTypes extends Component
         $uid = $event->tokenMatches[0];
         $data = $event->newValue;
 
-        $fieldUid = $data['field'];
-        $fieldId = Db::idByUid(Table::FIELDS, $fieldUid);
+        // Make sure the field has been synced
+        $fieldId = Db::idByUid(Table::FIELDS, $data['field']);
+        if ($fieldId === null) {
+            Craft::$app->getProjectConfig()->defer($event, [$this, __FUNCTION__]);
+            return;
+        }
 
-        $matrixBlockTypeUid = $data['matrixBlockType'];
-        $matrixBlockTypeId = Db::idByUid(Table::MATRIXBLOCKTYPES, $matrixBlockTypeUid);
+        // Make sure the matrix block type has been synced
+        $matrixBlockTypeId = Db::idByUid(Table::MATRIXBLOCKTYPES, $data['matrixBlockType']);
+        if ($matrixBlockTypeId === null) {
+            Craft::$app->getProjectConfig()->defer($event, [$this, __FUNCTION__]);
+            return;
+        }
 
         // Make sure fields and sites are processed
         ProjectConfigHelper::ensureAllSitesProcessed();
@@ -466,7 +474,7 @@ class BlockTypes extends Component
             'groupSortOrder',
             'sortOrder'
         ]));
-        
+
         // Use the fieldId to get the field and save the handle on to the model
         /** @var Field $matrixField */
         $matrixField = Craft::$app->fields->getFieldById($blockType->fieldId);